moved rrule hooks to pre('save') because:
Well those hooks are kinda a pain in the neck:
I suspect there is a bug when triggering post('validate') hooks of a
subdocument but could not wrap my head around it.

Anyway it works like this because although save hooks are not triggered
on insertMany() as we use in the between() function, the inserted
occurences have their rrule subdocument duplicated from the original.
That means that all its properties are processed correctly before being
inserted. (rrule.freq is uppercased and rrule.count is maxed when
rrule.until is present.)